Adds incrementing and decrementing of contexts
Adds commands for incrementing and decrementing contexts.

For headings, this command will cycle through "todo" keywords. The
default keyword values are the same as in `vscode-org-mode`, `TODO` and
`DONE`. Options to change these keywords are added to the configuration
settings.

For checkboxes, this command will check a checkbox, when incrementing,
and uncheck a checkbox when decrementing. Incrementing twice does
nothing. Similarly for decrementing.

Adds tests for incrementing and decrementing context.

This terminology of "contexts" comes from the `vscode-org-mode`
extension and possibly from emacs Org mode as well.